requirements.txt
streamlit
pdfplumber
pillow
requests
beautifulsoup4
anthropic
python-dotenv
reportlab
pytesseract